refactor(core): add and fix typescript
