--- packages/core/src/index.js
@@ -1,49 +1,0 @@
-export { $bridge, createHook, usePlugin } from './management';
-export * from './plugins/hook-manager'
-
-export {
-  ReactiveEffect,
-  activeSub,
-  resetTracking,
-  pauseTracking,
-  enableTracking,
-  shouldTrack,
-  startBatch,
-  endBatch,
-  stop,
-  onEffectCleanup,
-  refreshComputed,
-  EffectFlags
-} from '@vue-internals/reactivity/effect';
-
-export {
-  Dep,
-  Link,
-  track,
-  trigger,
-  getDepFromReactive,
-  globalVersion,
-  ARRAY_ITERATE_KEY,
-  ITERATE_KEY,
-  MAP_KEY_ITERATE_KEY
-} from '@vue-internals/reactivity/dep';
-
-export { getCurrentInstance, setCurrentInstance } from '@vue-internals/runtime-core/component';
-export {
-  queueJob,
-  queuePostFlushCb,
-  flushJobsUntil,
-  flushPostJobsUntil,
-  flushPreFlushCbs,
-  flushPostFlushCbs,
-  switchToAuto,
-  switchToManual,
-  toggleMode,
-  getJobAt,
-  nextTick,
-  endFlush,
-  SchedulerJobFlags,
-} from '@vue-internals/runtime-core/scheduler';
-
-export * from './lifecycle';
-export * from './conditional';
